WICKET-3690 Session is invalidated more than once because Session.sessionInvalidated is not reset
Moved the invocation of Session#detach() after the RequestCycleListener's #onDetach(). This way there is no need of RCL for the session detaching and it is executed always after the other listeners.
This way it looks like in 1.4 - there Session#detach() is called at the last step of RequestCycle#detach() - #threadDetach().

Unit tests pass.
Authentication and authorization examples work as expected.
Our application also works as expected even without null-ifying the 'sessionInvalidated' flag.



git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1101345 13f79535-47bb-0310-9956-ffa450edef68