Automate version numbering, post release message
The script has been expanded with determining the release version
automatically based on the version in the parent pom. The script will now
use that version in the script, but provide an escape hatch (CTRL-C to
abandon).

After building the release, the script provides post release steps that can
be taken, including upgrading master to a new development version, which is
not part of the automated script as the release needs to be validated by the
release manager first. There is no need to wait upgrading master to the new
development version once the release vote has started: it would impair new
development.

The script provides a list of copy/paste steps to perform after the release,
with correct values substituted for version numbers.