WICKET-4812 Make SerializationChecker easier for extending so custom checks can be added to it
Add support for exclusions by type before checking an object.
Re-work SessionSizeModel to be AbstractReadOnlyModel because otherwise it keeps a reference to the passed Session and during the calculation of the session size it fails that it is not detached. Now there is nothing to detach. Now it works with Session.get() only if it exists. It should not be possible to calculate the size of someone else's session anyway.