Manage experimental module versions from parent
Three Poms for the core project under the sky,
Seven for the Integrations in their XML of DOM,
Nine for Experimentals doomed to die,
One for the Maven parent POM,
In the Land of Maven where the Snapshots lie.
One POM to rule them all, One POM to find them,
One POM to bring them all and in the classpath bind them
In the Land of Maven where the Snapshots lie.