Fix for WICKET-4664
Added tests for RequestUtils.decodeParameters() : in order to be able to test it easily without the mocking or starting a complex requestCycle, extracted the static call to getCurrentCharset() as a method parameter. The extracted method is package visible only and visible for tests only.

When RequestUtils.decodeParameters() is called with a queryString beginning with a "?" it is automatically stripped and thus is not reflected into pageParameters.