WICKET-5523 Ajax indicator not working when display is set to none in css
Use jQuery#show() to show an element when an explicit display type is not requested. jQuery#show() has a logic to decide what is the most appropriate value to show a given HTMLElement (block or inline).
Ajax indicators are provided by the application and can be anything, so Wicket.DOM.showIncrementally() should not assume anything.