Fixed WICKET-5967 Unable to load i18n minified js
Moved location of minified (or not) resources to ResourceStreamLocator instead of the
resources themselves by adding a lookup for .min.<extension> before .<extension> in the list
of extensions to search for. This prioritizes the minified version before the non-minified version,
and allows for style_variation_locale resources to be located in minified versions as well.