WICKET-6254 Wicket WebSockets problem behind HTTP proxy without servlet context
Make it possible to override the parts needed to construct the WebSocket absolute url.
This is needed when the application runs in a servlet container that is proxied in front.
The constructed url has to point to the proxy server.
The application may use X-Forwarded-For-** headers set by the proxy, or custom configuration for different environments